requirements.txt
PyQt6
pygame
pyinstaller
pillow